+++ server/src/bruFiles/diagnostics/getAuthBlockSpecificDiagnostics.ts
@@ -3,17 +3,20 @@ import {
     isBlockDictionaryBlock,
     AuthBlockName,
     getMandatoryKeysForNonOAuth2Block,
-    ApiKeyAuthBlockKey,
+    ApiKeyAuthBlockKeys,
     ApiKeyAuthBlockPlacementValue,
     BooleanFieldValue,
     DictionaryBlock,
     OAuth2BlockCredentialsPlacementValue,
     OAuth2BlockTokenPlacementValue,
     OAuth2GrantType,
-    OAuth2ViaAuthorizationCodeBlockKey,
+    OAuth2ViaAuthorizationCodeBlockKeys,
     getMandatoryKeysForOAuth2Block,
     isDictionaryBlockSimpleField,
-    AuthBlockNameeExcludingOAuth2,
+    AuthBlockNamesExcludingOAuth2,
+    OAuth2AuthBlocksCommonKeys,
+    OAuth2BlockTokenSourceValue,
+    isAuthBlock,
 } from "@global_shared";
 import { checkNoDuplicateKeysAreDefinedForDictionaryBlock } from "./shared/checks/singleBlocks/checkNoDuplicateKeysAreDefinedForDictionaryBlock";
 import { checkNoKeysAreMissingForDictionaryBlock } from "./shared/checks/singleBlocks/checkNoKeysAreMissingForDictionaryBlock";
@@ -21,6 +24,7 @@ import { checkNoUnknownKeysAreDefinedInDictionaryBlock } from "./shared/checks/s
 import { checkValueForDictionaryBlockSimpleFieldIsValid } from "./shared/checks/singleBlocks/checkValueForDictionaryBlockSimpleFieldIsValid";
 import { DiagnosticWithCode } from "./interfaces";
 import { RelevantWithinAuthBlockDiagnosticCode } from "./shared/diagnosticCodes/relevantWithinAuthBlockDiagnosticCodeEnum";
+import { KnownDiagnosticCode } from "./shared/diagnosticCodes/knownDiagnosticCodeDefinition";
 
 export function getAuthBlockSpecificDiagnostics(
     filePath: string,
@@ -30,18 +34,16 @@ export function getAuthBlockSpecificDiagnostics(
         return [];
     }
 
-    const mandatoryKeys: string[] = [];
     const diagnostics: (DiagnosticWithCode | undefined)[] = [];
 
     if (
-        (Object.values(AuthBlockName) as string[]).includes(authBlock.name) &&
+        isAuthBlock(authBlock.name) &&
         authBlock.name != AuthBlockName.OAuth2Auth
     ) {
-        mandatoryKeys.push(
-            ...getMandatoryKeysForNonOAuth2Block(
-                authBlock.name as AuthBlockNameeExcludingOAuth2,
-            ),
+        const mandatoryKeys = getMandatoryKeysForNonOAuth2Block(
+            authBlock.name as AuthBlockNamesExcludingOAuth2,
         );
+
         diagnostics.push(
             checkNoKeysAreMissingForDictionaryBlock(
                 authBlock,
@@ -61,25 +63,19 @@ export function getAuthBlockSpecificDiagnostics(
             ) ?? []),
         );
 
-        if (
-            authBlock.name == AuthBlockName.ApiKeyAuth &&
-            authBlock.content.some(
-                ({ key }) => key == ApiKeyAuthBlockKey.Placement,
-            )
-        ) {
-            const field = authBlock.content.find(
-                ({ key }) => key == ApiKeyAuthBlockKey.Placement,
+        if (authBlock.name == AuthBlockName.ApiKeyAuth) {
+            diagnostics.push(
+                ...checkValuesForFields(authBlock, [
+                    {
+                        key: ApiKeyAuthBlockKeys.Placement,
+                        allowedValues: Object.values(
+                            ApiKeyAuthBlockPlacementValue,
+                        ),
+                        diagnosticCode:
+                            RelevantWithinAuthBlockDiagnosticCode.InvalidApiKeyAuthValueForPlacement,
+                    },
+                ]),
             );
-
-            if (field && isDictionaryBlockSimpleField(field)) {
-                diagnostics.push(
-                    checkValueForDictionaryBlockSimpleFieldIsValid(
-                        field,
-                        Object.values(ApiKeyAuthBlockPlacementValue),
-                        RelevantWithinAuthBlockDiagnosticCode.InvalidApiKeyAuthValueForPlacement,
-                    ),
-                );
-            }
         }
     } else if (authBlock.name == AuthBlockName.OAuth2Auth) {
         diagnostics.push(
@@ -97,7 +93,7 @@ function getDiagnosticsForOAuth2AuthBlock(
     const diagnostics: (DiagnosticWithCode | undefined)[] = [];
 
     const grantTypeFields = authBlock.content.filter(
-        ({ key }) => key == OAuth2ViaAuthorizationCodeBlockKey.GrantType,
+        ({ key }) => key == OAuth2ViaAuthorizationCodeBlockKeys.GrantType,
     );
 
     const diagnosticsForGrantTypeField: (DiagnosticWithCode | undefined)[] = [];
@@ -105,14 +101,14 @@ function getDiagnosticsForOAuth2AuthBlock(
     diagnosticsForGrantTypeField.push(
         checkNoKeysAreMissingForDictionaryBlock(
             authBlock,
-            [OAuth2ViaAuthorizationCodeBlockKey.GrantType],
+            [OAuth2ViaAuthorizationCodeBlockKeys.GrantType],
             RelevantWithinAuthBlockDiagnosticCode.KeysMissingInAuthBlock,
         ),
         ...(checkNoDuplicateKeysAreDefinedForDictionaryBlock(
             filePath,
             authBlock,
             RelevantWithinAuthBlockDiagnosticCode.DuplicateKeysDefinedInAuthBlock,
-            [OAuth2ViaAuthorizationCodeBlockKey.GrantType],
+            [OAuth2ViaAuthorizationCodeBlockKeys.GrantType],
         ) ?? []),
         grantTypeFields.length == 1 &&
             isDictionaryBlockSimpleField(grantTypeFields[0])
@@ -155,70 +151,26 @@ function getDiagnosticsForOAuth2AuthBlock(
 function checkValuesForOAuth2FieldsCommonForAllGrantTypes(
     authBlock: DictionaryBlock,
 ) {
-    const diagnostics: (DiagnosticWithCode | undefined)[] = [];
-
-    const credentialsPlacementFields = authBlock.content.filter(
-        ({ key }) =>
-            key == OAuth2ViaAuthorizationCodeBlockKey.CredentialsPlacement,
-    );
-
-    diagnostics.push(
-        credentialsPlacementFields.length == 1 &&
-            isDictionaryBlockSimpleField(credentialsPlacementFields[0])
-            ? checkValueForDictionaryBlockSimpleFieldIsValid(
-                  credentialsPlacementFields[0],
-                  Object.values(OAuth2BlockCredentialsPlacementValue),
-                  RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForCredentialsPlacement,
-              )
-            : undefined,
-    );
-
-    const tokenPlacementFields = authBlock.content.filter(
-        ({ key }) => key == OAuth2ViaAuthorizationCodeBlockKey.TokenPlacement,
-    );
-
-    diagnostics.push(
-        tokenPlacementFields.length == 1 &&
-            isDictionaryBlockSimpleField(tokenPlacementFields[0])
-            ? checkValueForDictionaryBlockSimpleFieldIsValid(
-                  tokenPlacementFields[0],
-                  Object.values(OAuth2BlockTokenPlacementValue),
-                  RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForTokenPlacement,
-              )
-            : undefined,
-    );
-
-    const autoFetchTokenFields = authBlock.content.filter(
-        ({ key }) => key == OAuth2ViaAuthorizationCodeBlockKey.AutoFetchToken,
-    );
-
-    diagnostics.push(
-        autoFetchTokenFields.length == 1 &&
-            isDictionaryBlockSimpleField(autoFetchTokenFields[0])
-            ? checkValueForDictionaryBlockSimpleFieldIsValid(
-                  autoFetchTokenFields[0],
-                  Object.values(BooleanFieldValue),
-                  RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForAutoFetchToken,
-              )
-            : undefined,
-    );
-
-    const autoRefreshTokenFields = authBlock.content.filter(
-        ({ key }) => key == OAuth2ViaAuthorizationCodeBlockKey.AutoRefreshToken,
-    );
-
-    diagnostics.push(
-        autoRefreshTokenFields.length == 1 &&
-            isDictionaryBlockSimpleField(autoRefreshTokenFields[0])
-            ? checkValueForDictionaryBlockSimpleFieldIsValid(
-                  autoRefreshTokenFields[0],
-                  Object.values(BooleanFieldValue),
-                  RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForAutoRefreshToken,
-              )
-            : undefined,
-    );
-
-    return diagnostics;
+    return checkValuesForFields(authBlock, [
+        {
+            key: OAuth2AuthBlocksCommonKeys.TokenPlacement,
+            allowedValues: Object.values(OAuth2BlockTokenPlacementValue),
+            diagnosticCode:
+                RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForTokenPlacement,
+        },
+        {
+            key: OAuth2AuthBlocksCommonKeys.TokenSource,
+            allowedValues: Object.values(OAuth2BlockTokenSourceValue),
+            diagnosticCode:
+                RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForTokenSource,
+        },
+        {
+            key: OAuth2ViaAuthorizationCodeBlockKeys.AutoFetchToken,
+            allowedValues: Object.values(BooleanFieldValue),
+            diagnosticCode:
+                RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForAutoFetchToken,
+        },
+    ]);
 }
 
 function checkValuesForOAuth2FieldsDependingOnGrantType(
@@ -250,21 +202,62 @@ function checkValuesForOAuth2FieldsDependingOnGrantType(
     );
 
     if (grantType == OAuth2GrantType.AuthorizationCode) {
-        const pkceFields = authBlock.content.filter(
-            ({ key }) => key == OAuth2ViaAuthorizationCodeBlockKey.Pkce,
-        );
-
         diagnostics.push(
-            pkceFields.length == 1 &&
-                isDictionaryBlockSimpleField(pkceFields[0])
-                ? checkValueForDictionaryBlockSimpleFieldIsValid(
-                      pkceFields[0],
-                      Object.values(BooleanFieldValue),
-                      RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForPkce,
-                  )
-                : undefined,
+            ...checkValuesForFields(authBlock, [
+                {
+                    key: OAuth2ViaAuthorizationCodeBlockKeys.Pkce,
+                    allowedValues: Object.values(BooleanFieldValue),
+                    diagnosticCode:
+                        RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForPkce,
+                },
+            ]),
+        );
+    }
+
+    if (grantType != OAuth2GrantType.Implicit) {
+        diagnostics.push(
+            ...checkValuesForFields(authBlock, [
+                {
+                    key: OAuth2ViaAuthorizationCodeBlockKeys.CredentialsPlacement,
+                    allowedValues: Object.values(
+                        OAuth2BlockCredentialsPlacementValue,
+                    ),
+                    diagnosticCode:
+                        RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForCredentialsPlacement,
+                },
+                {
+                    key: OAuth2ViaAuthorizationCodeBlockKeys.AutoRefreshToken,
+                    allowedValues: Object.values(BooleanFieldValue),
+                    diagnosticCode:
+                        RelevantWithinAuthBlockDiagnosticCode.InvalidOAuth2ValueForAutoRefreshToken,
+                },
+            ]),
         );
     }
 
     return diagnostics;
+}
+
+function checkValuesForFields(
+    authBlock: DictionaryBlock,
+    toCheck: {
+        key: string;
+        allowedValues: string[];
+        diagnosticCode: KnownDiagnosticCode;
+    }[],
+): (DiagnosticWithCode | undefined)[] {
+    return toCheck.map(({ key: keyToCheck, allowedValues, diagnosticCode }) => {
+        const matchingFields = authBlock.content.filter(
+            ({ key }) => key == keyToCheck,
+        );
+
+        return matchingFields.length == 1 &&
+            isDictionaryBlockSimpleField(matchingFields[0])
+            ? checkValueForDictionaryBlockSimpleFieldIsValid(
+                  matchingFields[0],
+                  allowedValues,
+                  diagnosticCode,
+              )
+            : undefined;
+    });
 }