+++ shared/index.ts
@@ -133,14 +133,15 @@ export * from "./languageUtils/commonBlocks/generic/variables/getBlocksWithoutVa
 export * from "./languageUtils/commonBlocks/generic/variables/getVariableNameForPositionInNonCodeBlock";
 
 export * from "./languageUtils/commonBlocks/authBlocks/authBlockNameEnum";
-export * from "./languageUtils/commonBlocks/authBlocks/authBlocksKeyEnums";
-export * from "./languageUtils/commonBlocks/authBlocks/oAuth2BlockValuesEnums";
+export * from "./languageUtils/commonBlocks/authBlocks/authBlocksKeyInterfaces";
+export * from "./languageUtils/commonBlocks/authBlocks/oAuth2GrantTypeEnum";
 export * from "./languageUtils/commonBlocks/authBlocks/apiKeyAuthBlockValuesEnums";
-export * from "./languageUtils/commonBlocks/authBlocks/oAuth2BlockCommonFieldsValuesEnums";
+export * from "./languageUtils/commonBlocks/authBlocks/oAuth2BlockCommonFieldsValues";
 export * from "./languageUtils/commonBlocks/authBlocks/getMandatoryKeysForNonOAuth2Block";
 export * from "./languageUtils/commonBlocks/authBlocks/getMandatoryKeysForOAuth2Block";
 export * from "./languageUtils/commonBlocks/authBlocks/getAuthTypeFromBlockName";
 export * from "./languageUtils/commonBlocks/authBlocks/isAuthBlock";
+export * from "./languageUtils/commonBlocks/authBlocks/additionalParamsBlocks/oauth2AdditionalParamsBlockNames";
 
 export * from "./languageUtils/commonBlocks/authModeBlock/authModeBlockKeyEnum";