+++ shared/languageUtils/commonBlocks/authBlocks/oAuth2BlockCommonFieldsValuesEnums.ts
@@ -1,9 +1,0 @@
-export enum OAuth2BlockCredentialsPlacementValue {
-    Body = "body",
-    BasicAuthHeader = "basic_auth_header",
-}
-
-export enum OAuth2BlockTokenPlacementValue {
-    Header = "header",
-    Url = "url",
-}