+++ shared/languageUtils/folderSettingsFiles/getValidBlockNames.ts
@@ -1,5 +1,6 @@
 import {
     AuthBlockName,
+    Oauth2AdditionalParamsBlockNames,
     RequestFileBlockName,
     SettingsFileSpecificBlock,
 } from "../..";
@@ -10,6 +11,7 @@ export function getValidBlockNames(): string[] {
         RequestFileBlockName.Headers,
         SettingsFileSpecificBlock.AuthMode,
         ...Object.values(AuthBlockName),
+        ...Object.values(Oauth2AdditionalParamsBlockNames),
         RequestFileBlockName.PreRequestVars,
         RequestFileBlockName.PostResponseVars,
         RequestFileBlockName.PreRequestScript,