+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - Improve grouping and sorting of context menu items in collection explorer.
 - Provide completions for keys in all dictionary blocks, except for auth blocks of type OAuth2.
+- Add support for blocks for additional params for OAuth2 auth (diagnostics, snippets).
+- Add support for new fields/values for OAuth2 auth blocks (grant type: new value 'implicit' and new field 'token_source').
 
 ## [0.22.1] (2026-04-02)